Modularized sns script
